+++ src/internal/adapter/WebEventForwarder.ts
@@ -6,6 +6,7 @@ import type {
   CastStateChangeEvent,
   ChromecastErrorEvent,
   ChromelessPlayer,
+  ContentProtectionErrorEvent as NativeContentProtectionErrorEvent,
   CurrentSourceChangeEvent as NativeCurrentSourceChangeEvent,
   DimensionChangeEvent as NativeDimensionChangeEvent,
   DurationChangeEvent as NativeDurationChangeEvent,
@@ -52,6 +53,7 @@ import {
   DefaultAirplayStateChangeEvent,
   DefaultChromecastChangeEvent,
   DefaultChromecastErrorEvent,
+  DefaultContentProtectionErrorEvent,
   DefaultCurrentSourceChangeEvent,
   DefaultDimensionChangeEvent,
   DefaultDurationChangeEvent,
@@ -112,6 +114,7 @@ export class WebEventForwarder {
     this._player.addEventListener('ratechange', this.onPlaybackRateChange);
     this._player.addEventListener('segmentnotfound', this.onSegmentNotFound);
     this._player.addEventListener('volumechange', this.onVolumeChangeEvent);
+    this._player.addEventListener('contentprotectionerror', this.onContentProtectionError);
     this._player.addEventListener('dimensionchange', this.onDimensionChange);
     this._player.addEventListener('resize', this.onVideoResize);
 
@@ -163,6 +166,7 @@ export class WebEventForwarder {
     this._player.removeEventListener('ratechange', this.onPlaybackRateChange);
     this._player.removeEventListener('segmentnotfound', this.onSegmentNotFound);
     this._player.removeEventListener('volumechange', this.onVolumeChangeEvent);
+    this._player.removeEventListener('contentprotectionerror', this.onContentProtectionError);
     this._player.removeEventListener('dimensionchange', this.onDimensionChange);
     this._player.removeEventListener('resize', this.onVideoResize);
 
@@ -223,6 +227,21 @@ export class WebEventForwarder {
       new DefaultErrorEvent({
         errorCode: event.errorObject.code.toString(),
         errorMessage: event.errorObject.message,
+      }),
+    );
+  };
+
+  private readonly onContentProtectionError = (event: NativeContentProtectionErrorEvent) => {
+    const { code, message: errorMessage, url, status, statusText, response, systemCode } = event.errorObject;
+    this._facade.dispatchEvent(
+      new DefaultContentProtectionErrorEvent({
+        errorCode: code.toString(),
+        errorMessage,
+        url,
+        status,
+        statusText,
+        response,
+        systemCode,
       }),
     );
   };